chore: configurable logger
